WIP: 프리셋 API 연동 관련 구현 중(테스트 필요)
