+++ .gitignore
@@ -43,3 +43,7 @@ service_key.json
 *.bk
 /bk/*
 */test-page/*
+
+# test
+*.test.tsx
+*/test/*
