+++ package.json
@@ -8,6 +8,8 @@
     "build": "next build --webpack",
     "start": "next start",
     "lint": "next lint",
+    "test": "vitest run",
+    "test:watch": "vitest",
     "analyze": "ANALYZE=true npm run build",
     "build:analyze": "ANALYZE=true npm run build"
   },
@@ -55,15 +57,19 @@
     "@babel/preset-typescript": "^7.28.5",
     "@next/bundle-analyzer": "^16.0.4",
     "@svgr/webpack": "^8.1.0",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.2",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "babel-loader": "^10.0.0",
     "eslint": "^9.39.2",
     "eslint-config-next": "^16.1.6",
+    "jsdom": "^28.1.0",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "typescript": "^5",
+    "vitest": "^4.0.18"
   },
   "browserslist": [
     "last 2 versions",
