+++ src/app/seller/mypage/preset/page.tsx
@@ -6,7 +6,10 @@ import {
   ExtraPolicyPicker,
   ExtraPolicySection,
   PageHeader,
-} from "@/components/features/mypage/preset/PresetSections";
+} from "@/components/features/mypage/preset/sections";
+import { usePresetPage } from "@/hooks/seller/preset/usePresetPage";
+import { usePresetQueries } from "@/hooks/seller/preset/usePresetQueries";
+import { ConfirmDialog } from "@/components/features/dashboard/SubmitConfirmDialog";
 import { useState } from "react";
 
 
@@ -17,74 +20,67 @@ import { useState } from "react";
  * - 추가 정책 설정 버튼 클릭 시 아래에 새로운 section 생성 [v]
  * - section 생성 시, 시간대 설정 부분 추가해야 함 [v]
  * 기능 구현 
- *   - name 값 설정 
- *   - 값 유효성 검사
- *   - 값 저장 (API 연동)
+ *   - name 값 설정 [v]
+ *   - 값 유효성 검사 [v]
+ *   - 값 저장 (API 연동) [진행중]
  * 
  */
 
-type ExtraPolicyType = "PEAK" | "CLOSE";
-
-interface ExtraPolicySectionState {
-  id: string;
-  type: ExtraPolicyType;
-  slotIds: string[];
-}
-
 function PresetPage() {
-  const [extraPolicies, setExtraPolicies] = useState<ExtraPolicySectionState[]>([]);
-  const [policyCounter, setPolicyCounter] = useState(0);
-  const [slotCounter, setSlotCounter] = useState(0);
+  const { extraPolicies, hiddenTypes, addPolicy, getPolicyTitle, 
+    handlePresetCancel, isPresetSavedDialogOpen, setIsPresetSavedDialogOpen } = usePresetPage();
+  const [pendingSubmitAction, setPendingSubmitAction] = useState<(() => void) | null>(null);
 
-  const addPolicy = (type: ExtraPolicyType) => {
-    if (extraPolicies.some((policy) => policy.type === type)) return;
-
-    const nextPolicyId = policyCounter + 1;
-    const nextSlotId = slotCounter + 1;
-    setPolicyCounter(nextPolicyId);
-    setSlotCounter(nextSlotId);
-
-    setExtraPolicies((prev) => [
-      ...prev,
-      {
-        id: `policy-${nextPolicyId}`,
-        type,
-        slotIds: [`slot-${nextSlotId}`],
-      },
-    ]);
+  const handleBaseSaveRequest = (submitAction: () => void) => {
+    setPendingSubmitAction(() => submitAction);
+    setIsPresetSavedDialogOpen(true);
   };
 
-  const addSlot = (policyId: string) => {
-    const nextSlotId = slotCounter + 1;
-    setSlotCounter(nextSlotId);
-    setExtraPolicies((prev) =>
-      prev.map((policy) =>
-        policy.id === policyId ? { ...policy, slotIds: [...policy.slotIds, `slot-${nextSlotId}`] } : policy
-      )
-    );
+  const handlePresetConfirmSubmit = () => {
+    pendingSubmitAction?.();
+    setPendingSubmitAction(null);
+    setIsPresetSavedDialogOpen(false);
   };
 
-  const getPolicyTitle = (type: ExtraPolicyType) => (type === "PEAK" ? "피크타임 노쇼 정책" : "마감 정책");
-  const hiddenTypes = Array.from(new Set(extraPolicies.map((policy) => policy.type)));
+  const handlePresetCancelSubmit = () => {
+    setPendingSubmitAction(null);
+    handlePresetCancel();
+  };
 
+  const { presets, loading, error } = usePresetQueries();
+  if (loading) return <div className='w-screen h-full flex items-center justify-center'>로딩 중...</div>;
+  if (error) return <div className='w-screen h-full flex items-center justify-center'>에러 발생: {error}</div>;
   return (
-    <DashBoardLayout>
-      <div className="max-w-[1400px] w-full h-full flex flex-col gap-40 pt-40 pb-20">
-        <PageHeader />
-        <BasePolicySection />
+    <>
+      <DashBoardLayout>
+        <div className="max-w-[1400px] w-full h-full flex flex-col gap-40 pt-40 pb-20">
+          <PageHeader />
+          <BasePolicySection presetData={presets[0]} onSaveRequest={handleBaseSaveRequest} />
 
-        {extraPolicies.map((policy) => (
-          <ExtraPolicySection
-            key={policy.id}
-            title={getPolicyTitle(policy.type)}
-            slotIds={policy.slotIds}
-            // onAddSlot={() => addSlot(policy.id)}
-          />
-        ))}
+          {extraPolicies.map((policy) => (
+            <ExtraPolicySection
+              key={policy.id}
+              title={getPolicyTitle(policy.type)}
+              slotIds={policy.slotIds}
+              onSaveRequest={handleBaseSaveRequest}
+              // onAddSlot={() => addSlot(policy.id)}
+            />
+          ))}
 
-        <ExtraPolicyPicker onAdd={addPolicy} hiddenTypes={hiddenTypes} />
-      </div>
-    </DashBoardLayout>
+          <ExtraPolicyPicker onAdd={addPolicy} hiddenTypes={hiddenTypes} />
+        </div>
+      </DashBoardLayout>
+      
+      {/* 저장 버튼 클릭 시 묻는 다이얼로그 */}
+      <ConfirmDialog
+        open = {isPresetSavedDialogOpen}
+        onOpenChange = {setIsPresetSavedDialogOpen}
+        onCancel={handlePresetCancelSubmit}
+        onConfirm={handlePresetConfirmSubmit} 
+        title = '프리셋 저장'
+        description = '프리셋을 저장하시겠습니까?'
+      />
+    </>
   );
 }
