+++ src/components/features/mypage/preset/sections/PageHeader.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+export function PageHeader() {
+  return (
+    <div className="flex flex-col gap-20">
+      <div className="title7 font-foreground-normal">노쇼 프리셋</div>
+      <div className="body3 text-foreground-secondary">시스템이 자동으로 적용할 노쇼 판매 규칙을 설정합니다.</div>
+    </div>
+  );
+}
