+++ src/components/features/mypage/preset/sections/index.ts
@@ -1,0 +1,4 @@
+export { PageHeader } from "./PageHeader";
+export { BasePolicySection } from "./BasePolicySection";
+export { ExtraPolicySection } from "./ExtraPolicySection";
+export { ExtraPolicyPicker } from "./ExtraPolicyPicker";
