+++ src/hooks/seller/preset/usePresetForm.ts
@@ -2,6 +2,8 @@
 
 import { useEffect, useMemo, useState } from "react";
 import { z } from "zod";
+import { PresetData, usePresetUpdate } from "./usePresetQueries";
+import { usePresetPage } from "./usePresetPage";
 
 export type TimePeriod = "AM" | "PM";
 
@@ -18,7 +20,7 @@ export type TimeSlot = {
 
 export type PolicyFormValues = {
   name: string;
-  discountRate: number;
+  discountPercent: number;
   visitAvailableHour: number;
   visitAvailableMinute: number;
   waitingMinutes: number;
@@ -37,8 +39,8 @@ export const createDefaultTimeSlot = (): TimeSlot => ({
 
 const timeValueSchema = z.object({
   period: z.enum(["AM", "PM"]),
-  hour: z.number().min(1).max(12),
-  minute: z.number().min(0).max(59),
+  hour: z.coerce.number().min(1).max(12),
+  minute: z.coerce.number().min(0).max(59),
 });
 
 const timeSlotSchema = z.object({
@@ -49,10 +51,10 @@ const timeSlotSchema = z.object({
 const createPolicyFormSchema = (requireTimeSlots: boolean) =>
   z.object({
     name: z.string().trim().min(1, "정책 이름을 입력하세요."),
-    discountRate: z.number().min(30, "할인율을 입력하세요.(30 이상)").max(90, "할인율은 최대 90%입니다."),
-    visitAvailableHour: z.number().min(0, "방문 가능 시간을 입력하세요.").max(5,"방문 가능 시간은 최대 5시간입니다."),
-    visitAvailableMinute: z.number().min(1, "방문 가능 분을 입력하세요.").max(59, "방문 가능 분은 최대 59분입니다."),
-    waitingMinutes: z.number().min(1, "판매 대기 시간을 입력하세요.").max(300, "판매 대기 시간은 최대 300분입니다."),
+    discountPercent: z.coerce.number().min(30, "할인율을 입력하세요.(30 이상)").max(90, "할인율은 최대 90%입니다."),
+    visitAvailableHour: z.coerce.number().min(0, "방문 가능 시간을 입력하세요.").max(5,"방문 가능 시간은 최대 5시간입니다."),
+    visitAvailableMinute: z.coerce.number().min(1, "방문 가능 분을 입력하세요.").max(59, "방문 가능 분은 최대 59분입니다."),
+    waitingMinutes: z.coerce.number().min(1, "판매 대기 시간을 입력하세요.").max(300, "판매 대기 시간은 최대 300분입니다."),
     timeSlots: requireTimeSlots ? z.array(timeSlotSchema).min(1, "시간대를 1개 이상 설정하세요.") : z.array(timeSlotSchema),
   });
 
@@ -61,21 +63,37 @@ type UsePresetFormOptions = {
   sectionLabel: string;
   requireTimeSlots: boolean;
   slotIds?: string[];
+  presetData?: PresetData;
 };
 
-export function usePresetForm({ initialValues, sectionLabel, requireTimeSlots, slotIds }: UsePresetFormOptions) {
-  const [values, setValues] = useState<PolicyFormValues>(initialValues);
+const normalizePolicyFormValues = (values: PolicyFormValues): PolicyFormValues => ({
+  ...values,
+  discountPercent: Number(values.discountPercent),
+  visitAvailableHour: Number(values.visitAvailableHour),
+  visitAvailableMinute: Number(values.visitAvailableMinute),
+  waitingMinutes: Number(values.waitingMinutes),
+});
+
+export function usePresetForm({ initialValues, sectionLabel, requireTimeSlots, slotIds, presetData }: UsePresetFormOptions) {
+  // const { handlePresetCancel, handlePresetConfirm } = usePresetPage();
+  const [values, setValues] = useState<PolicyFormValues>(normalizePolicyFormValues(initialValues));
   const [errors, setErrors] = useState<PolicyFormErrors>({});
   const schema = useMemo(() => createPolicyFormSchema(requireTimeSlots), [requireTimeSlots]);
+  const presetUpdate = usePresetUpdate(presetData!);
 
   const mapZodFieldErrors = (fieldErrors: Record<string, string[] | undefined>): PolicyFormErrors => ({
     name: fieldErrors.name?.[0],
-    discountRate: fieldErrors.discountRate?.[0],
+    discountPercent: fieldErrors.discountPercent?.[0] ?? fieldErrors.discountRate?.[0],
     visitAvailableHour: fieldErrors.visitAvailableHour?.[0],
     visitAvailableMinute: fieldErrors.visitAvailableMinute?.[0],
     waitingMinutes: fieldErrors.waitingMinutes?.[0],
     timeSlots: fieldErrors.timeSlots?.[0],
   });
+
+  useEffect(() => {
+    setValues(normalizePolicyFormValues(initialValues));
+    setErrors({});
+  }, [initialValues]);
 
   useEffect(() => {
     if (!slotIds) return;
@@ -89,8 +107,23 @@ export function usePresetForm({ initialValues, sectionLabel, requireTimeSlots, s
     });
   }, [slotIds]);
 
+  const numericFields: Array<keyof Omit<PolicyFormValues, "timeSlots">> = [
+    "discountPercent",
+    "visitAvailableHour",
+    "visitAvailableMinute",
+    "waitingMinutes",
+  ];
+
+  const isNumericField = (field: keyof Omit<PolicyFormValues, "timeSlots">) => numericFields.includes(field);
+
   const handleFieldChange = (field: keyof Omit<PolicyFormValues, "timeSlots">, value: string | number) => {
-    setValues((prev) => ({ ...prev, [field]: value }));
+    const normalizedValue = isNumericField(field)
+      ? typeof value === "number"
+        ? value
+        : Number(value)
+      : value;
+
+    setValues((prev) => ({ ...prev, [field]: normalizedValue }));
     setErrors((prev) => ({ ...prev, [field]: undefined }));
   };
 
@@ -118,16 +151,19 @@ export function usePresetForm({ initialValues, sectionLabel, requireTimeSlots, s
     // TODO: parsed data API 연동 전처리 후 저장 로직 구현
     const preprocessedData = dataPreprocessing(parsed.data);
     console.log(`${sectionLabel} 전처리된 데이터`, preprocessedData);
+    if(!preprocessedData) return;
+    if(!presetUpdate) return;
+    presetUpdate.mutate(preprocessedData);
   };
 
-  const dataPreprocessing = (data: PolicyFormValues) => {
+  const dataPreprocessing = (data: PolicyFormValues): PresetData | undefined => {
     const visitAvailableMinutes = data.visitAvailableHour * 60 + data.visitAvailableMinute;
     if(visitAvailableMinutes > 300) {
       console.error("방문 가능 시간은 최대 5시간(300분)입니다.");
       alert("방문 가능 시간은 최대 5시간(300분)입니다.");
       return;
     }
-    const discountPercent = data.discountRate;
+    const discountPercent = data.discountPercent;
     const saleDelayMinutes = data.waitingMinutes  ;
 
       if(discountPercent < 30 || discountPercent > 90) {
@@ -149,11 +185,13 @@ export function usePresetForm({ initialValues, sectionLabel, requireTimeSlots, s
     if(dataName === "기본 노쇼 정책") dataName = "기본";
 
     return {
+      presetId: presetData?.presetId ?? "",
       name: dataName,
       discountPercent,
       visitAvailableMinutes,
       saleDelayMinutes,
-    }
+      updatedAt: presetData?.updatedAt ?? new Date().toISOString(),
+    };
   }
 
   return {
